Move fingertree implementation into subdirectory.
