+++ src/main/java/org/seanpatrickmiller/containers/fingertree/SequenceFactory.java
@@ -1,16 +1,16 @@
 package org.seanpatrickmiller.containers.fingertree;
 
-import org.seanpatrickmiller.containers.fingertree.Empty;
-import org.seanpatrickmiller.containers.fingertree.FingerTree;
-import org.seanpatrickmiller.containers.fingertree.Measured;
-import org.seanpatrickmiller.containers.fingertree.Single;
+import org.seanpatrickmiller.containers.fingertree.impl.Empty;
+import org.seanpatrickmiller.containers.fingertree.impl.FingerTree;
+import org.seanpatrickmiller.containers.fingertree.impl.Measured;
+import org.seanpatrickmiller.containers.fingertree.impl.Single;
 import org.seanpatrickmiller.containers.util.Func;
 import org.seanpatrickmiller.containers.util.Identity;
 import org.seanpatrickmiller.containers.util.Monoids;
 
 /**
  * Factory for creating indexed sequences backed by
- * {@link org.seanpatrickmiller.containers.fingertree.FingerTree}.
+ * {@link org.seanpatrickmiller.containers.fingertree.impl.FingerTree}.
  */
 public final class SequenceFactory<A>
 {
