+++ src/main/java/org/seanpatrickmiller/containers/fingertree/impl/Four.java
@@ -1,20 +1,22 @@
-package org.seanpatrickmiller.containers.fingertree;
+package org.seanpatrickmiller.containers.fingertree.impl;
 
 import org.seanpatrickmiller.containers.util.Func;
 
-final class Three<V, A> extends Digit<V, A>
+final class Four<V, A> extends Digit<V, A>
 {
     final A a;
     final A b;
     final A c;
+    final A d;
 
-    Three(final Measured<V, A> m, final A a, final A b, final A c)
+    Four(final Measured<V, A> m, final A a, final A b, final A c, final A d)
     {
         super(m);
 
         this.a = a;
         this.b = b;
         this.c = c;
+        this.d = d;
     }
 
     @Override
@@ -25,8 +27,8 @@ final class Three<V, A> extends Digit<V, A>
     {
         return new Deep<V, A>(
             m,
-            new Four<V, A>(m, x, a, b, c),
-            mid,
+            new Two<V, A>(m, x, a),
+            mid.pushFront(new Node3<V, A>(m, b, c, d)),
             right);
     }
 
@@ -39,38 +41,38 @@ final class Three<V, A> extends Digit<V, A>
         return new Deep<V, A>(
             m,
             left,
-            mid,
-            new Four<V, A>(m, a, b, c, x));
+            mid.pushBack(new Node3<V, A>(m, a, b, c)),
+            new Two<V, A>(m, d, x));
     }
 
     @Override
     Digit<V, A> reverse(final Func<A, A> f)
     {
-        return new Three<V, A>(m, f.call(c), f.call(b), f.call(a));
+        return new Four<V, A>(m, f.call(d), f.call(c), f.call(b), f.call(a));
     }
 
     @Override
     <B> Digit<V, B> map(final Func<A, B> f, final Measured<V, B> m)
     {
-        return new Three<V, B>(m, f.call(a), f.call(b), f.call(c));
+        return new Four<V, B>(m, f.call(a), f.call(b), f.call(c), f.call(d));
     }
 
     @Override
     Digit<V, A> map(final Func<A, A> f)
     {
-        return new Three<V, A>(m, f.call(a), f.call(b), f.call(c));
+        return new Four<V, A>(m, f.call(a), f.call(b), f.call(c), f.call(d));
     }
 
     @Override
     <B> B foldRight(final Func<A, Func<B, B>> f, final B zero)
     {
-        return f.call(a).call(f.call(b).call(f.call(c).call(zero)));
+        return f.call(a).call(f.call(b).call(f.call(c).call(f.call(d).call(zero))));
     }
 
     @Override
     <B> B foldLeft(final Func<B, Func<A, B>> f, final B zero)
     {
-        return f.call(f.call(f.call(zero).call(a)).call(b)).call(c);
+        return f.call(f.call(f.call(f.call(zero).call(a)).call(b)).call(c)).call(d);
     }
 
     @Override
@@ -82,7 +84,7 @@ final class Three<V, A> extends Digit<V, A>
             return new Split<Digit<V, A>, A>(
                 null,
                 a,
-                new Two<V, A>(m, b, c));
+                new Three<V, A>(m, b, c, d));
         }
 
         final V vab = m.sum(va, m.measure(b));
@@ -91,12 +93,21 @@ final class Three<V, A> extends Digit<V, A>
             return new Split<Digit<V, A>, A>(
                 new One<V, A>(m, a),
                 b,
-                new One<V, A>(m, c));
+                new Two<V, A>(m, c, d));
+        }
+
+        final V vabc = m.sum(vab, m.measure(c));
+        if (pred.call(vabc))
+        {
+            return new Split<Digit<V, A>, A>(
+                new Two<V, A>(m, a, b),
+                c,
+                new One<V, A>(m, d));
         }
 
         return new Split<Digit<V, A>, A>(
-            new Two<V, A>(m, a, b),
-            c,
+            new Three<V, A>(m, a, b, c),
+            d,
             null);
     }
 
@@ -107,7 +118,7 @@ final class Three<V, A> extends Digit<V, A>
             m,
             new Two<V, A>(m, a, b),
             new Empty<V, Node<V, A>>(m.nodeMeasured()),
-            new One<V, A>(m, c));
+            new Two<V, A>(m, c, d));
     }
 
     @Override
@@ -119,24 +130,24 @@ final class Three<V, A> extends Digit<V, A>
     @Override
     Digit<V, A> tail()
     {
-        return new Two<V, A>(m, b, c);
+        return new Three<V, A>(m, b, c, d);
     }
 
     @Override
     A rhead()
     {
-        return c;
+        return d;
     }
 
     @Override
     Digit<V, A> rtail()
     {
-        return new Two<V, A>(m, a, b);
+        return new Three<V, A>(m, a, b, c);
     }
 
     @Override
     public java.lang.String toString()
     {
-        return "Three(" + a + "," + b + "," + c + ")";
+        return "Four(" + a + "," + b + "," + c + "," + d + ")";
     }
 }
