+++ src/main/java/org/seanpatrickmiller/containers/fingertree/impl/One.java
@@ -1,18 +1,16 @@
-package org.seanpatrickmiller.containers.fingertree;
+package org.seanpatrickmiller.containers.fingertree.impl;
 
 import org.seanpatrickmiller.containers.util.Func;
 
-final class Two<V, A> extends Digit<V, A>
+final class One<V, A> extends Digit<V, A>
 {
     final A a;
-    final A b;
 
-    Two(final Measured<V, A> m, final A a, final A b)
+    One(final Measured<V, A> m, final A a)
     {
         super(m);
 
         this.a = a;
-        this.b = b;
     }
 
     @Override
@@ -23,7 +21,7 @@ final class Two<V, A> extends Digit<V, A>
     {
         return new Deep<V, A>(
             m,
-            new Three<V, A>(m, x, a, b),
+            new Two<V, A>(m, x, a),
             mid,
             right);
     }
@@ -38,65 +36,49 @@ final class Two<V, A> extends Digit<V, A>
             m,
             left,
             mid,
-            new Three<V, A>(m, a, b, x));
+            new Two<V, A>(m, a, x));
     }
 
     @Override
     Digit<V, A> reverse(final Func<A, A> f)
     {
-        return new Two<V, A>(m, f.call(b), f.call(a));
+        return new One<V, A>(m, f.call(a));
     }
 
     @Override
     <B> Digit<V, B> map(final Func<A, B> f, final Measured<V, B> m)
     {
-        return new Two<V, B>(m, f.call(a), f.call(b));
+        return new One<V, B>(m, f.call(a));
     }
 
     @Override
-    Digit<V, A> map(final Func<A,A> f)
+    Digit<V, A> map(final Func<A, A> f)
     {
-        return new Two<V, A>(m, f.call(a), f.call(b));
+        return new One<V, A>(m, f.call(a));
     }
 
     @Override
     <B> B foldRight(final Func<A, Func<B, B>> f, final B zero)
     {
-        return f.call(a).call(f.call(b).call(zero));
+        return f.call(a).call(zero);
     }
 
     @Override
     <B> B foldLeft(final Func<B, Func<A, B>> f, final B zero)
     {
-        return f.call(f.call(zero).call(a)).call(b);
+        return f.call(zero).call(a);
     }
 
     @Override
     Split<Digit<V, A>, A> split(final Func<V, Boolean> pred, final V measure)
     {
-        final V va = m.sum(measure, m.measure(a));
-        if (pred.call(va))
-        {
-            return new Split<Digit<V, A>, A>(
-                null,
-                a,
-                new One<V, A>(m, b));
-        }
-
-        return new Split<Digit<V, A>, A>(
-            new One<V, A>(m, a),
-            b,
-            null);
+        return new Split<Digit<V, A>, A>(null, a, null);
     }
 
     @Override
     FingerTree<V, A> toTree()
     {
-        return new Deep<V, A>(
-            m,
-            new One<V, A>(m, a),
-            new Empty<V, Node<V, A>>(m.nodeMeasured()),
-            new One<V, A>(m, b));
+        return new Single<V, A>(m, a);
     }
 
     @Override
@@ -108,24 +90,26 @@ final class Two<V, A> extends Digit<V, A>
     @Override
     Digit<V, A> tail()
     {
-        return new One<V, A>(m, b);
+        throw new java.lang.UnsupportedOperationException(
+            "Digit.tail(): cannot call on instance of One");
     }
 
     @Override
     A rhead()
     {
-        return b;
+        return a;
     }
 
     @Override
     Digit<V, A> rtail()
     {
-        return new One<V, A>(m, a);
+        throw new java.lang.UnsupportedOperationException(
+            "Digit.rtail(): cannot call on instance of One");
     }
 
     @Override
     public java.lang.String toString()
     {
-        return "Two(" + a + "," + b + ")";
+        return "One(" + a + ")";
     }
 }
