+++ src/main/java/org/seanpatrickmiller/containers/fingertree/impl/Three.java
@@ -1,22 +1,20 @@
-package org.seanpatrickmiller.containers.fingertree;
+package org.seanpatrickmiller.containers.fingertree.impl;
 
 import org.seanpatrickmiller.containers.util.Func;
 
-final class Four<V, A> extends Digit<V, A>
+final class Three<V, A> extends Digit<V, A>
 {
     final A a;
     final A b;
     final A c;
-    final A d;
 
-    Four(final Measured<V, A> m, final A a, final A b, final A c, final A d)
+    Three(final Measured<V, A> m, final A a, final A b, final A c)
     {
         super(m);
 
         this.a = a;
         this.b = b;
         this.c = c;
-        this.d = d;
     }
 
     @Override
@@ -27,8 +25,8 @@ final class Four<V, A> extends Digit<V, A>
     {
         return new Deep<V, A>(
             m,
-            new Two<V, A>(m, x, a),
-            mid.pushFront(new Node3<V, A>(m, b, c, d)),
+            new Four<V, A>(m, x, a, b, c),
+            mid,
             right);
     }
 
@@ -41,38 +39,38 @@ final class Four<V, A> extends Digit<V, A>
         return new Deep<V, A>(
             m,
             left,
-            mid.pushBack(new Node3<V, A>(m, a, b, c)),
-            new Two<V, A>(m, d, x));
+            mid,
+            new Four<V, A>(m, a, b, c, x));
     }
 
     @Override
     Digit<V, A> reverse(final Func<A, A> f)
     {
-        return new Four<V, A>(m, f.call(d), f.call(c), f.call(b), f.call(a));
+        return new Three<V, A>(m, f.call(c), f.call(b), f.call(a));
     }
 
     @Override
     <B> Digit<V, B> map(final Func<A, B> f, final Measured<V, B> m)
     {
-        return new Four<V, B>(m, f.call(a), f.call(b), f.call(c), f.call(d));
+        return new Three<V, B>(m, f.call(a), f.call(b), f.call(c));
     }
 
     @Override
     Digit<V, A> map(final Func<A, A> f)
     {
-        return new Four<V, A>(m, f.call(a), f.call(b), f.call(c), f.call(d));
+        return new Three<V, A>(m, f.call(a), f.call(b), f.call(c));
     }
 
     @Override
     <B> B foldRight(final Func<A, Func<B, B>> f, final B zero)
     {
-        return f.call(a).call(f.call(b).call(f.call(c).call(f.call(d).call(zero))));
+        return f.call(a).call(f.call(b).call(f.call(c).call(zero)));
     }
 
     @Override
     <B> B foldLeft(final Func<B, Func<A, B>> f, final B zero)
     {
-        return f.call(f.call(f.call(f.call(zero).call(a)).call(b)).call(c)).call(d);
+        return f.call(f.call(f.call(zero).call(a)).call(b)).call(c);
     }
 
     @Override
@@ -84,7 +82,7 @@ final class Four<V, A> extends Digit<V, A>
             return new Split<Digit<V, A>, A>(
                 null,
                 a,
-                new Three<V, A>(m, b, c, d));
+                new Two<V, A>(m, b, c));
         }
 
         final V vab = m.sum(va, m.measure(b));
@@ -93,21 +91,12 @@ final class Four<V, A> extends Digit<V, A>
             return new Split<Digit<V, A>, A>(
                 new One<V, A>(m, a),
                 b,
-                new Two<V, A>(m, c, d));
-        }
-
-        final V vabc = m.sum(vab, m.measure(c));
-        if (pred.call(vabc))
-        {
-            return new Split<Digit<V, A>, A>(
-                new Two<V, A>(m, a, b),
-                c,
-                new One<V, A>(m, d));
+                new One<V, A>(m, c));
         }
 
         return new Split<Digit<V, A>, A>(
-            new Three<V, A>(m, a, b, c),
-            d,
+            new Two<V, A>(m, a, b),
+            c,
             null);
     }
 
@@ -118,7 +107,7 @@ final class Four<V, A> extends Digit<V, A>
             m,
             new Two<V, A>(m, a, b),
             new Empty<V, Node<V, A>>(m.nodeMeasured()),
-            new Two<V, A>(m, c, d));
+            new One<V, A>(m, c));
     }
 
     @Override
@@ -130,24 +119,24 @@ final class Four<V, A> extends Digit<V, A>
     @Override
     Digit<V, A> tail()
     {
-        return new Three<V, A>(m, b, c, d);
+        return new Two<V, A>(m, b, c);
     }
 
     @Override
     A rhead()
     {
-        return d;
+        return c;
     }
 
     @Override
     Digit<V, A> rtail()
     {
-        return new Three<V, A>(m, a, b, c);
+        return new Two<V, A>(m, a, b);
     }
 
     @Override
     public java.lang.String toString()
     {
-        return "Four(" + a + "," + b + "," + c + "," + d + ")";
+        return "Three(" + a + "," + b + "," + c + ")";
     }
 }
